Module robusta_jni::convert::unchecked

Infallible conversion traits.

These traits allow for a leaner generated glue code, with possibly some performance benefits.

These conversion traits can be enabled to be used during code generation with the unchecked option on the call_type attribute, as so:

#[call_type(unchecked)]

These functions will panic should any conversion fail.

Traits

FromJavaValue

Conversion trait from Java values to Rust values, analogous to From. Used when converting types that are input to JNI-available functions.

IntoJavaValue

Conversion trait from Rust values to Java values, analogous to Into. Used when converting types returned from JNI-available functions.

Derive Macros

FromJavaValue

IntoJavaValue