Trait robusta_jni::convert::unchecked::FromJavaValue

pub trait FromJavaValue<'env: 'borrow, 'borrow>: Signature {
    type Source: JavaValue<'env>;

    const SIG_TYPE: &'static str;

    fn from(s: Self::Source, env: &'borrow JNIEnv<'env>) -> Self;
}

Conversion trait from Java values to Rust values, analogous to From. Used when converting types that are input to JNI-available functions.

Notes on derive macro

The same notes on TryFromJavaValue apply.

Associated Types

type Source: JavaValue<'env>

Conversion source type.

Associated Constants

const SIG_TYPE: &'static str

Signature of the target type. By default, use the one defined on the Signature trait for the implementing type.

Required methods

fn from(s: Self::Source, env: &'borrow JNIEnv<'env>) -> Self

Perform the conversion.